+++ src/materials/transmissive.h
@@ -10,7 +10,7 @@ public:
 
 	//Constructor
 	Transmissive() { refractive_index = 0.0; }
-	Transmissive(float refractive_index);
+	Transmissive(double refractive_index);
 
 	//Inherited methods
 	Vector3D getReflectance(const Vector3D& n, const Vector3D& wo, const Vector3D& wi) const;
